Added tests for expo-dev-launcher AppDelegate plugin
