+++ packages/expo-dev-launcher/plugin/src/withDevLauncher.ts
@@ -8,6 +8,7 @@ import {
 import { ExpoConfig } from '@expo/config-types';
 import fs from 'fs';
 import path from 'path';
+import { withDevLauncherAppDelegate } from './withDevLauncherAppDelegate';
 
 const DEV_LAUNCHER_ANDROID_IMPORT = 'expo.modules.devlauncher.DevLauncherController';
 const DEV_LAUNCHER_ON_NEW_INTENT = `
@@ -24,50 +25,6 @@ const DEV_LAUNCHER_ANDROID_INIT = 'DevLauncherController.initialize(this, getRea
 
 const DEV_LAUNCHER_POD_IMPORT =
   "pod 'expo-dev-menu', path: '../node_modules/expo-dev-menu', :configurations => :debug";
-const DEV_LAUNCHER_APP_DELEGATE_SOURCE_FOR_URL = `  #if defined(EX_DEV_LAUNCHER_ENABLED)
-return [[EXDevLauncherController sharedInstance] sourceUrl];
-#else
-return [[RCTBundleURLProvider sharedSettings] jsBundleURLForBundleRoot:@"index" fallbackResource:nil];
-#endif`;
-const DEV_LAUNCHER_APP_DELEGATE_ON_DEEP_LINK = `#if defined(EX_DEV_LAUNCHER_ENABLED)
-if ([EXDevLauncherController.sharedInstance onDeepLink:url options:options]) {
-return true;
-}
-#endif
-return [RCTLinkingManager application:application openURL:url options:options];`;
-const DEV_LAUNCHER_APP_DELEGATE_IOS_IMPORT = `
-#if defined(EX_DEV_LAUNCHER_ENABLED)
-#include <EXDevLauncher/EXDevLauncherController.h>
-#endif`;
-const DEV_LAUNCHER_APP_DELEGATE_CONTROLLER_DELEGATE = `
-#if defined(EX_DEV_LAUNCHER_ENABLED)
-@implementation AppDelegate (EXDevLauncherControllerDelegate)
-
-- (void)devLauncherController:(EXDevLauncherController *)developmentClientController
-      didStartWithSuccess:(BOOL)success
-{
-developmentClientController.appBridge = [self initializeReactNativeApp];
-EXSplashScreenService *splashScreenService = (EXSplashScreenService *)[UMModuleRegistryProvider getSingletonModuleForClass:[EXSplashScreenService class]];
-[splashScreenService showSplashScreenFor:self.window.rootViewController];
-}
-
-@end
-#endif
-`;
-const DEV_LAUNCHER_APP_DELEGATE_INIT = `#if defined(EX_DEV_LAUNCHER_ENABLED)
-      EXDevLauncherController *contoller = [EXDevLauncherController sharedInstance];
-      [contoller startWithWindow:self.window delegate:self launchOptions:launchOptions];
-    #else
-      [self initializeReactNativeApp];
-    #endif`;
-
-const DEV_LAUNCHER_APP_DELEGATE_BRIDGE = `#if defined(EX_DEV_LAUNCHER_ENABLED)
-  NSDictionary *launchOptions = [EXDevLauncherController.sharedInstance getLaunchOptions];
-#else
-  NSDictionary *launchOptions = self.launchOptions;
-#endif
-
-  RCTBridge *bridge = [[RCTBridge alloc] initWithDelegate:self launchOptions:launchOptions];`;
 
 async function readFileAsync(path: string): Promise<string> {
   return fs.promises.readFile(path, 'utf8');
@@ -123,7 +80,7 @@ async function editMainApplication(
     return await saveFileAsync(mainApplicationPath, mainApplication);
   } catch (e) {
     WarningAggregator.addWarningIOS(
-      'ios-devMenu',
+      'expo-dev-launcher',
       `Couldn't modified MainApplication.java - ${e}.`
     );
   }
@@ -136,25 +93,7 @@ async function editPodfile(config: ExportedConfigWithProps, action: (podfile: st
 
     return await saveFileAsync(podfilePath, podfile);
   } catch (e) {
-    WarningAggregator.addWarningIOS('ios-devMenu', `Couldn't modified AppDelegate.m - ${e}.`);
-  }
-}
-
-async function editAppDelegate(
-  config: ExportedConfigWithProps,
-  action: (appDelegate: string) => string
-) {
-  const appDelegatePath = path.join(
-    config.modRequest.platformProjectRoot,
-    config.modRequest.projectName!,
-    'AppDelegate.m'
-  );
-
-  try {
-    const appDelegate = action(await readFileAsync(appDelegatePath));
-    return await saveFileAsync(appDelegatePath, appDelegate);
-  } catch (e) {
-    WarningAggregator.addWarningIOS('ios-devMenu', `Couldn't modified AppDelegate.m - ${e}.`);
+    WarningAggregator.addWarningIOS('expo-dev-launcher', `Couldn't modified AppDelegate.m - ${e}.`);
   }
 }
 
@@ -203,7 +142,7 @@ const withDevLauncherActivity: ConfigPlugin = config => {
       config.modResults.contents = content;
     } else {
       WarningAggregator.addWarningAndroid(
-        'android-devLauncher',
+        'expo-dev-launcher',
         `Cannot automatically configure MainActivity if it's not java`
       );
     }
@@ -235,57 +174,6 @@ const withDevLauncherPodfile: ConfigPlugin = config => {
   ]);
 };
 
-const withDevLauncherAppDelegate: ConfigPlugin = config => {
-  return withDangerousMod(config, [
-    'ios',
-    async config => {
-      await editAppDelegate(config, appDelegate => {
-        if (!appDelegate.includes(DEV_LAUNCHER_APP_DELEGATE_IOS_IMPORT)) {
-          const lines = appDelegate.split('\n');
-          lines.splice(1, 0, DEV_LAUNCHER_APP_DELEGATE_IOS_IMPORT);
-
-          appDelegate = lines.join('\n');
-        }
-
-        if (!appDelegate.includes(DEV_LAUNCHER_APP_DELEGATE_INIT)) {
-          appDelegate = appDelegate.replace(
-            /(didFinishLaunchingWithOptions([^}])*)\[self initializeReactNativeApp\];(([^}])*})/,
-            `$1${DEV_LAUNCHER_APP_DELEGATE_INIT}$3`
-          );
-        }
-
-        if (!appDelegate.includes(DEV_LAUNCHER_APP_DELEGATE_BRIDGE)) {
-          appDelegate = appDelegate.replace(
-            'RCTBridge *bridge = [[RCTBridge alloc] initWithDelegate:self launchOptions:self.launchOptions];',
-            DEV_LAUNCHER_APP_DELEGATE_BRIDGE
-          );
-        }
-
-        if (!appDelegate.includes(DEV_LAUNCHER_APP_DELEGATE_SOURCE_FOR_URL)) {
-          appDelegate = appDelegate.replace(
-            'return [[RCTBundleURLProvider sharedSettings] jsBundleURLForBundleRoot:@"index" fallbackResource:nil];',
-            DEV_LAUNCHER_APP_DELEGATE_SOURCE_FOR_URL
-          );
-        }
-
-        if (!appDelegate.includes(DEV_LAUNCHER_APP_DELEGATE_ON_DEEP_LINK)) {
-          appDelegate = appDelegate.replace(
-            'return [RCTLinkingManager application:application openURL:url options:options];',
-            DEV_LAUNCHER_APP_DELEGATE_ON_DEEP_LINK
-          );
-        }
-
-        if (!appDelegate.includes(DEV_LAUNCHER_APP_DELEGATE_CONTROLLER_DELEGATE)) {
-          appDelegate += DEV_LAUNCHER_APP_DELEGATE_CONTROLLER_DELEGATE;
-        }
-
-        return appDelegate;
-      });
-
-      return config;
-    },
-  ]);
-};
 const withDevLauncher = (config: ExpoConfig) => {
   config = withDevLauncherActivity(config);
   config = withDevLauncherApplication(config);
